Converts floats/doubles to decimals
We don't really need the precision of the other data types (certainly not planning on displaying values > 2dp) so might as well keep everything consistent

+++ MediatonicApi/Models/Animal.cs
@@ -5,8 +5,8 @@ namespace MediatonicApi.Models
     {
         public uint Id { get; set; }
         public string TypeName { get; set; }
-        public double HungerPerSecond { get; set; }
-        public double SadnessPerSecond { get; set; }
+        public decimal HungerPerSecond { get; set; }
+        public decimal SadnessPerSecond { get; set; }
 
         public uint MaxHunger { get; set; }
         public int MinHappiness { get; set; }
